Declutter Particles and introduce ThinParticleSystem
